# [julia_wrapper] Slightly improve performance