This PR slightly improves compilation time with Julia by moving the code to a function, and marking the help text as a constant, which are the very basic [performance tips](https://docs.julialang.org/en/v1/manual/performance-tips/) recommended when writing Julia code.  Note that most of the changes involve indentation, so the diff is easier to read by ignoring whitespace changes.

With this change, on my machine (Intel i7-4870HQ CPU) I get a runtime improvement in the range 10%-40%, depending on the julia version used.  It isn't a huge saving, but I guess shaving off a few hundreds milliseconds isn't too bad :slightly_smiling_face: 

<details><summary>Details of benchmarks</summary>
<p>

`input.jl` file:
```julia
function axpy(a, x, y)
    @simd for idx in eachindex(x, y)
        @inbounds y[idx] = muladd(a, x[idx], y[idx])
    end
    return nothing
end

precompile(axpy, (Float32, Vector{Float32}, Vector{Float32}))
```

Benchmark (`julia_wrapper.jl` is the script in this PR, `julia_wrapper_main.jl` is [the script currently in the `main` branch](https://github.com/compiler-explorer/compiler-explorer/blob/0be4fbf9c55c77c18babacfce4dcc86a1dee844f/etc/scripts/julia_wrapper.jl), I used [juliaup](https://github.com/JuliaLang/juliaup) to launch the different julia versions)

```console
% for version in 1.7 1.8 1.9 1.10; do hyperfine --warmup=1 "julia +${version} --startup-file=no julia_wrapper.jl input.jl output.ll --format=llvm" "julia +${version} --startup-file=no julia_wrapper_main.jl input.jl output.ll --format=llvm"; done
Benchmark 1: julia +1.7 --startup-file=no --project=. julia_wrapper.jl input.jl output.ll --format=llvm
  Time (mean ± σ):      1.364 s ±  0.035 s    [User: 1.512 s, System: 0.511 s]
  Range (min … max):    1.299 s …  1.399 s    10 runs
 
Benchmark 2: julia +1.7 --startup-file=no --project=. julia_wrapper_main.jl input.jl output.ll --format=llvm
  Time (mean ± σ):      1.899 s ±  0.010 s    [User: 2.072 s, System: 0.541 s]
  Range (min … max):    1.878 s …  1.914 s    10 runs
 
Summary
  julia +1.7 --startup-file=no --project=. julia_wrapper.jl input.jl output.ll --format=llvm ran
    1.39 ± 0.04 times faster than julia +1.7 --startup-file=no --project=. julia_wrapper_main.jl input.jl output.ll --format=llvm
Benchmark 1: julia +1.8 --startup-file=no --project=. julia_wrapper.jl input.jl output.ll --format=llvm
  Time (mean ± σ):      1.702 s ±  0.048 s    [User: 1.848 s, System: 0.535 s]
  Range (min … max):    1.642 s …  1.784 s    10 runs
 
Benchmark 2: julia +1.8 --startup-file=no --project=. julia_wrapper_main.jl input.jl output.ll --format=llvm
  Time (mean ± σ):      1.858 s ±  0.053 s    [User: 1.995 s, System: 0.558 s]
  Range (min … max):    1.792 s …  1.956 s    10 runs
 
Summary
  julia +1.8 --startup-file=no --project=. julia_wrapper.jl input.jl output.ll --format=llvm ran
    1.09 ± 0.04 times faster than julia +1.8 --startup-file=no --project=. julia_wrapper_main.jl input.jl output.ll --format=llvm
Benchmark 1: julia +1.9 --startup-file=no --project=. julia_wrapper.jl input.jl output.ll --format=llvm
  Time (mean ± σ):      1.693 s ±  0.059 s    [User: 1.750 s, System: 0.256 s]
  Range (min … max):    1.616 s …  1.776 s    10 runs
 
Benchmark 2: julia +1.9 --startup-file=no --project=. julia_wrapper_main.jl input.jl output.ll --format=llvm
  Time (mean ± σ):      2.338 s ±  0.035 s    [User: 2.391 s, System: 0.253 s]
  Range (min … max):    2.284 s …  2.389 s    10 runs
 
Summary
  julia +1.9 --startup-file=no --project=. julia_wrapper.jl input.jl output.ll --format=llvm ran
    1.38 ± 0.05 times faster than julia +1.9 --startup-file=no --project=. julia_wrapper_main.jl input.jl output.ll --format=llvm
Benchmark 1: julia +1.10 --startup-file=no --project=. julia_wrapper.jl input.jl output.ll --format=llvm
  Time (mean ± σ):      1.541 s ±  0.024 s    [User: 1.596 s, System: 0.256 s]
  Range (min … max):    1.501 s …  1.580 s    10 runs
 
Benchmark 2: julia +1.10 --startup-file=no --project=. julia_wrapper_main.jl input.jl output.ll --format=llvm
  Time (mean ± σ):      1.752 s ±  0.029 s    [User: 1.808 s, System: 0.255 s]
  Range (min … max):    1.707 s …  1.791 s    10 runs
 
Summary
  julia +1.10 --startup-file=no --project=. julia_wrapper.jl input.jl output.ll --format=llvm ran
    1.14 ± 0.03 times faster than julia +1.10 --startup-file=no --project=. julia_wrapper_main.jl input.jl output.ll --format=llvm
```

</p>
</details>